requirements.txt
albumentations
matplotlib
numpy
pyyaml
scikit-learn
segmentation-models-pytorch
tifffile
timm
torch
